ElCamino.Azure.Data.Tables refactor
