--- src/ElCamino.AspNetCore.Identity.AzureTable/Helpers/Odata/EdmType.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace ElCamino.AspNetCore.Identity.AzureTable.Helpers
-{
-    public enum EdmType
-    {
-        Binary,
-        Boolean,
-        DateTime,
-        Double,
-        Guid,        
-        Int32,
-        Int64,
-        String
-
-    }
-}
