+++ src/ElCamino.Azure.Data.Tables/EntityMapExtensions.cs
@@ -1,18 +1,12 @@
 ﻿// MIT License Copyright 2020 (c) David Melendez. All rights reserved. See License.txt in the project root for license information.
 
-using Azure.Data.Tables;
-using System;
 using System.Collections.Concurrent;
-using System.Collections.Generic;
-using System.Linq;
 using System.Reflection;
 using System.Runtime.Serialization;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace ElCamino.AspNetCore.Identity.AzureTable.Helpers
+namespace Azure.Data.Tables
 {
-    public static class EntityMapHelper
+    public static class EntityMapExtensions
     {
         private static readonly ConcurrentDictionary<string, PropertyInfo[]> TypeProperties = new();
