+++ src/ElCamino.Azure.Data.Tables/TableQuery.cs
@@ -1,11 +1,31 @@
-﻿using System;
+﻿// -----------------------------------------------------------------------------------------
+// <copyright file="TableQuery.cs" company="Microsoft">
+//    Copyright 2013 Microsoft Corporation
+// 
+//    Licensed under the Apache License, Version 2.0 (the "License");
+//    you may not use this file except in compliance with the License.
+//    You may obtain a copy of the License at
+//      http://www.apache.org/licenses/LICENSE-2.0
+// 
+//    Unless required by applicable law or agreed to in writing, software
+//    distributed under the License is distributed on an "AS IS" BASIS,
+//    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//    See the License for the specific language governing permissions and
+//    limitations under the License.
+// </copyright>
+// -----------------------------------------------------------------------------------------
+
+using System;
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.Globalization;
 using System.Text;
 
-namespace ElCamino.AspNetCore.Identity.AzureTable.Helpers
+namespace Azure.Data.Tables
 {
+    /// <summary>
+    /// From https://github.com/Azure/azure-storage-net/blob/v9.3.2/Lib/Common/Table/TableQuery.cs
+    /// </summary>
     public class TableQuery
     {
         public int? TakeCount { get; set; }
